+++ src/components/AdminRoute/index.js
@@ -1,0 +1,2 @@
+import AdminRoute from './AdminRoute';
+export default AdminRoute;
