+++ src/components/UserRoute/index.js
@@ -1,0 +1,2 @@
+import UserRoute from './UserRoute';
+export default UserRoute;
